Move MapVal and ArrayVal to IVal inner classes. + delete ValueSortedMap.java
